+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Tweak secondary view may not show from floating panel.
 
 ### [1.0.4] - 2023-08-29
 ### Fixed
-- Crash after tweak list scroll animation end on iOS 16
+- Crash after tweak list scroll animation end on iOS 16.
 
 ## [1.0.3] - 2022-11-16
 ### Fixed